+++ nfs/src/server/nfs40/op_set_clientid_confirm.rs
@@ -74,19 +74,17 @@ mod integration_tests {
         // setup clients
         let res_client1 = client1.execute(request).await;
         let (client1_id, client1_confirm) = match res_client1.result.unwrap() {
-            NfsResOp4::Opsetclientid(res) => match res {
-                SetClientId4res::Resok4(resok) => (resok.clientid, resok.setclientid_confirm),
-                _ => panic!("Unexpected response"),
-            },
+            NfsResOp4::Opsetclientid(SetClientId4res::Resok4(resok)) => {
+                (resok.clientid, resok.setclientid_confirm)
+            }
             _ => panic!("Unexpected response"),
         };
 
         let res_client2 = client2.execute(res_client1.request).await;
         let (client2_id, client2_confirm) = match res_client2.result.unwrap() {
-            NfsResOp4::Opsetclientid(res) => match res {
-                SetClientId4res::Resok4(resok) => (resok.clientid, resok.setclientid_confirm),
-                _ => panic!("Unexpected response"),
-            },
+            NfsResOp4::Opsetclientid(SetClientId4res::Resok4(resok)) => {
+                (resok.clientid, resok.setclientid_confirm)
+            }
             _ => panic!("Unexpected response"),
         };